+++ Sources/Swift/Integrations/SessionReplay/RRWeb/SentryRRWebOptionsEvent.swift
@@ -3,13 +3,19 @@ import Foundation
 
 @objc class SentryRRWebOptionsEvent: SentryRRWebCustomEvent {
     
+    init(timestamp: Date, customOptions: [String: Any]) {
+        super.init(timestamp: timestamp, tag: "options", payload: customOptions)
+    }
+    
     init(timestamp: Date, options: SentryReplayOptions) {
         var payload: [String: Any] = [
             "sessionSampleRate": options.sessionSampleRate,
             "errorSampleRate": options.onErrorSampleRate,
             "maskAllText": options.maskAllText,
             "maskAllImages": options.maskAllImages,
-            "quality": String(describing: options.quality)
+            "quality": String(describing: options.quality),
+            "nativeSdkName": SentryMeta.sdkName,
+            "nativeSdkVersion": SentryMeta.versionString
         ]
         
         if !options.maskedViewClasses.isEmpty {