+++ Sources/Swift/Integrations/SessionReplay/SentrySessionReplay.swift
@@ -41,6 +41,7 @@ class SentrySessionReplay: NSObject {
     private let touchTracker: SentryTouchTracker?
     private let dispatchQueue: SentryDispatchQueueWrapper
     private let lock = NSLock()
+    var replayTags: [String: Any]?
     
     var isRunning: Bool {
         displayLink.isRunning()
@@ -248,6 +249,7 @@ class SentrySessionReplay: NSObject {
     private func captureSegment(segment: Int, video: SentryVideoInfo, replayId: SentryId, replayType: SentryReplayType) {
         let replayEvent = SentryReplayEvent(eventId: replayId, replayStartTimestamp: video.start, replayType: replayType, segmentId: segment)
         
+        replayEvent.sdk = self.replayOptions.sdkInfo
         replayEvent.timestamp = video.end
         replayEvent.urls = video.screens
         
@@ -260,7 +262,11 @@ class SentrySessionReplay: NSObject {
         }
         
         if segment == 0 {
-            events.append(SentryRRWebOptionsEvent(timestamp: video.start, options: self.replayOptions))
+            if let customOptions = replayTags {
+                events.append(SentryRRWebOptionsEvent(timestamp: video.start, customOptions: customOptions))
+            } else {
+                events.append(SentryRRWebOptionsEvent(timestamp: video.start, options: self.replayOptions))
+            }
         }
         
         let recording = SentryReplayRecording(segmentId: segment, video: video, extraEvents: events)