+++ Sources/Swift/Integrations/UserFeedback/SentryUserFeedbackWidget.swift
@@ -19,9 +19,11 @@ struct SentryUserFeedbackWidget {
             
             lazy var button = SentryUserFeedbackWidgetButtonView(config: config, action: { _ in
                 self.setWidget(visible: false)
-                let form = SentryUserFeedbackForm(config: self.config, delegate: self)
+                let form = SentryUserFeedbackFormController(config: self.config, delegate: self)
                 form.presentationController?.delegate = self
-                self.present(form, animated: self.config.animations)
+                self.present(form, animated: self.config.animations) {  
+                    self.config.onFormOpen?()
+                }
             })
             
             let config: SentryUserFeedbackConfiguration
@@ -75,7 +77,9 @@ struct SentryUserFeedbackWidget {
             
             func closeForm() {
                 setWidget(visible: true)
-                dismiss(animated: config.animations)
+                dismiss(animated: config.animations, completion: {
+                    self.config.onFormClose?()
+                })
             }
             
             // MARK: SentryUserFeedbackFormDelegate
@@ -92,6 +96,7 @@ struct SentryUserFeedbackWidget {
             
             func presentationControllerDidDismiss(_ presentationController: UIPresentationController) {
                 setWidget(visible: true)
+                self.config.onFormClose?()
             }
         }